backend/requirements.txt
# FastAPI and server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
pydantic>=2.10.0

# Vector database
qdrant-client>=1.12.0

# LLM
google-generativeai>=0.8.0

# Embeddings
sentence-transformers>=3.3.0

# Database
asyncpg>=0.30.0
psycopg2-binary>=2.9.10

# HTTP client
httpx>=0.28.0
